+++ djangojs/tests/test_context.py
@@ -226,6 +226,21 @@ class ContextTestMixin(object):
         self.assertNotIn('user', result)
         self.assertTrue(len(result.keys()) > 0)
 
+    @override_settings(JS_CONTEXT=['STATIC_URL', 'LANGUAGE_CODE'])
+    def test_context_whitelist(self):
+        '''Should only include context variable from settings.JS_CONTEXT if set'''
+        result = self.process_request()
+        self.assertEqual(len(result.keys()), 2 + 1) #: User is always in context
+        self.assertIn('STATIC_URL', result)
+        self.assertIn('LANGUAGE_CODE', result)
+
+    @override_settings(JS_CONTEXT_EXCLUDE=['STATIC_URL', 'LANGUAGE_CODE'])
+    def test_context_blacklist(self):
+        '''Should exclude context variable from settings.JS_CONTEXT_EXCLUDE if set'''
+        result = self.process_request()
+        self.assertNotIn('STATIC_URL', result)
+        self.assertNotIn('LANGUAGE_CODE', result)
+
 
 @override_settings(
     TEMPLATE_CONTEXT_PROCESSORS=TEST_CONTEXT_PROCESSORS,